리워드 double 형으로 처리

+++ myworkchainserver/src/main/java/bankware/finlab/myworkchain/server/dto/SendRewardRequest.java
@@ -14,5 +14,5 @@ public class SendRewardRequest {
 	private String receiverAddress;
 	
 	//리워드 양
-	private int valueAmount;
+	private double valueAmount;
 }
